Move CardResult into .cardpayments module.
