+++ CardPayments/src/main/java/com/paypal/android/cardpayments/ApproveOrderListener.kt
@@ -1,7 +1,6 @@
 package com.paypal.android.cardpayments
 
 import androidx.annotation.MainThread
-import com.paypal.android.cardpayments.model.CardResult
 import com.paypal.android.corepayments.PayPalSDKError
 
 /**
