+++ PayPalWebPayments/src/main/java/com/paypal/android/paypalwebpayments/PayPalDeepLinkUrlResult.kt
@@ -4,10 +4,7 @@ import android.net.Uri
 import androidx.annotation.NonNull
 import org.json.JSONObject
 
-internal class PayPalDeepLinkUrlResult constructor(
-    @NonNull uri: Uri,
-    @NonNull metadata: JSONObject
-) {
+internal class PayPalDeepLinkUrlResult constructor(uri: Uri, metadata: JSONObject) {
     val payerId: String? = uri.getQueryParameter("PayerID")
     val intent: String? = uri.getQueryParameter("intent")
     val opType: String? = uri.getQueryParameter("opType")
